First pass: refactoring the 'Selection' sidebar
