--- src/pages/annotate/SidebarSection/CurrentSelection/CurrentSelectionMetadata/index.ts
@@ -1,1 +1,0 @@
-export * from './CurrentSelectionMetadata';
