+++ src/pages/annotate/SidebarSection/CurrentSelection/PropertiesForm/PropertyKeys.ts
@@ -16,7 +16,7 @@ export const createSafeKeys = (tuples: BodyAndEntityType[]) => {
   // Converts schema names to keys we can safely use with formik, and
   // keeps a lookup table (or, rather, list): key -> { name, body }
   const safeKeys: [string, { name: string, body: W3CAnnotationBody }][] = tuples.reduce((safeKeys, { body, entityType }) => (
-    [...safeKeys, ...entityType.properties.map(property => ([toSafeKey(body, property.name), { name: property.name, body }]))]
+    [...safeKeys, ...(entityType.properties || []).map(property => ([toSafeKey(body, property.name), { name: property.name, body }]))]
   ), []);
 
   const getKey = (b: W3CAnnotationBody, n: string) => {
