+++ src/pages/annotate/SidebarSection/CurrentSelection/PropertiesForm/index.ts
@@ -1,0 +1,1 @@
+export * from './PropertiesForm';
