+++ src/pages/annotate/SidebarSection/CurrentSelection/PropertiesFormSection/index.ts
@@ -1,0 +1,1 @@
+export * from './PropertiesFormSection';
